requirements.txt
Flask==2.3.3
flask-sqlalchemy==3.1.1
SQLAlchemy==2.0.20
python-dotenv==1.0.0
gunicorn==21.2.0
gevent==23.9.1
psycopg2-binary==2.9.7  
alembic==1.12.0  